requirements.txt
# Image Tampering Detection Dependencies
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
requests>=2.31.0
tqdm>=4.65.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
joblib>=1.3.0
scipy>=1.10.0
timm>=0.9.0

# API Dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
